// src/Components/Model/iMovieID.tsx
import iBundle from "./iBundles";
import iDate from "./iDate"
import iSession from "./iSesstion";
import { iTicket } from "./iTicket";

export interface iMovieID {
    name: string;
    subName: string;
    age: number;
    url: string;
    id: string;
    cinemaId: string;
    scheduledFilmId: string;
    sessionId: string;
    areaCategoryCodes: any[];
    isAllocatedSeating: boolean;
    allowChildAdmits: boolean;
    seatsAvailable: number;
    allowComplimentaryTickets: boolean;
    eventId: string;
    priceGroupCode: string;
    screenName: string;
    screenNameAlt: string;
    screenNumber: number;
    cinemaOperatorCode: string;
    formatCode: string;
    formatHopk: string;
    salesChannels: string;
    sessionAttributesNames: any[];
    conceptAttributesNames: any[];
    allowTicketSales: boolean;
    hasDynamicallyPricedTicketsAvailable: boolean;
    playThroughId?: any;
    sessionBusinessDate: Date;
    sessionDisplayPriority: number;
    groupSessionsByAttribute: boolean;
    soldoutStatus: number;
    typeCode: string;
    dayOfWeekLabel: string;
    dayOfWeekKey: string;
    showDate: string;
    showTime: string;
    caption: string;
    code: string;
    version: string;
    sessions: Array<iSession>;
    bundles: Array<iBundle>;
    slug: string;
    vistaName: string;
    description: string;
    order: number;
    phone: string;
    address: string;
    cityId: string;
    mapEmbeb: string;
    status: number;
    reward: number;
    createdAt: Date;
    updatedAt: Date;
    createdBy?: any;
    updatedBy: string;
    oldId: number;
    imageUrls: string[];
    ticket: Array<iTicket>;
    imageLandscape: string;
    imagePortrait?: any;
    longitude: string;
    latitude: string;
    cityIds?: any;
    dates: Array<iDate>;
}